% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sound_DMR.R
\name{add_changepoint_info}
\alias{add_changepoint_info}
\title{Store Changepoint Information}
\usage{
add_changepoint_info(whole_df, changepoint, col_name)
}
\arguments{
\item{whole_df}{(df) - all the information for one gene}

\item{changepoint}{(df) - the changepoint summary}

\item{col_name}{(str) - the name of the column upon which the changepoint analysis
is being conducted}
}
\value{
whole_df (df) - information for one gene including the
changepoint summary
}
\description{
Add the important changepoint summary info to the dataframe from
which the changepoints were found
}
