% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sound_DMR.R, R/Sound_DMR_formatted.R
\name{add_zoom_coords}
\alias{add_zoom_coords}
\title{add_zoom_coords}
\usage{
add_zoom_coords(
  target,
  gene_cord_df,
  geneco_index,
  gcoord_exist = TRUE,
  Gene_col = "Gene_name"
)

add_zoom_coords(
  target,
  gene_cord_df,
  geneco_index,
  gcoord_exist = TRUE,
  Gene_col = "Gene_name"
)
}
\arguments{
\item{target}{(df) - Subset of ONT-methyl bed to positions pertaining to a single gene at a time}

\item{geneco_index}{(int) - This is index that refers to the row number of the geneco df, required to assign proper Alphabet codes.}

\item{gcoord_exist}{(Boolean) - This is to use the function only the gene_cord_df file has the location of the gene.}

\item{Gene_col}{(str) - Use this column to specifify wether to add Gene Names or Ids in the Zoomframe.}
}
\value{
Zoomframe (df) - Similar to Megaframe except this includes more information on targets, positions zero'ed to ATG for each target and a few other information with an additional column that included zoom codes

Zoomframe (df) - Similar to Megaframe except this includes more information on targets, positions zero'ed to ATG for each target and a few other information with an additional column that included zoom codes
}
\description{
A function to add Zoom codes based on the gene positions.
Currently the codes for this is as below
1- Anything that is only between gene start and gene stop
2- Anything that is between Adaptive start and Adaptive Stop
0- Anything that doesn't fall within in the above - to ensure we don't include these in the DMR analysis.

A function to add Zoom codes based on the gene positions.
Currently the codes for this is as below
1- Anything that is only between gene start and gene stop
2- Anything that is between Adaptive start and Adaptive Stop
0- Anything that doesn't fall within in the above - to ensure we don't include these in the DMR analysis.
}
