% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sound_DMR.R, R/Sound_DMR_formatted.R
\name{boot_score}
\alias{boot_score}
\title{Boot_score}
\usage{
boot_score(
  sound_score_obj = NA,
  target_gene = NA,
  target_start = -1000,
  target_end = 0,
  nboots = 1000,
  scoring_col_name = "dmr_score",
  direction_DMR = "positive"
)

boot_score(
  sound_score_obj = NA,
  target_gene = NA,
  target_start = -1000,
  target_end = 0,
  nboots = 1000,
  scoring_col_name = "dmr_score",
  direction_DMR = "positive"
)
}
\arguments{
\item{sound_score_obj}{(list) - Aggregated Changepoint Object created from sound_score}

\item{target_gene}{(str) - name of the target gene that was hit with oligos}

\item{target_start}{(str) - Start position relative to gene start that was hit with oligos}

\item{target_end}{(str) - Stop position relative to gene start that was hit with oligos}

\item{nboots}{(str) - Number of bootstrap replicates}

\item{scoring_col_name}{(str) - Name of the score column to run bootstrapping on (one of dmr_score or dmr_score2)}
}
\value{
Boot_Obj (list) - File that contains information updated on target region info and bootstrap values

Boot_Obj (list) - File that contains information updated on target region info and bootstrap values
}
\description{
Bootstrap analysis comparing target region with other regions of genome

Bootstrap analysis comparing target region with other regions of genome
}
