% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sound_DMR.R, R/Sound_DMR_formatted.R
\name{changepoint_analysis}
\alias{changepoint_analysis}
\title{Changepoint Analysis}
\usage{
changepoint_analysis(
  whole_df,
  CG_penalty = int,
  CHG_penalty = int,
  CHH_penalty = int,
  target_genes = c(),
  save_plots = FALSE,
  z_col = "column"
)

changepoint_analysis(
  whole_df,
  CG_penalty = int,
  CHG_penalty = int,
  CHH_penalty = int,
  target_genes = c(),
  save_plots = FALSE,
  z_col = "column"
)
}
\arguments{
\item{whole_df}{(df) - the information such as genes, z scores, etc}

\item{CG_penalty, CHG_penalty, CHH_penalty}{(int) - penalty values for the
change point analysis. The higher the value the fewer changepoints that will
be created. The lower the value the more changepoints that will be created.}

\item{target_genes}{(list of strings) - a list of the target genes}

\item{save_plots}{(boolean) - if TRUE save plots to working directory, default
is FALSE}

\item{z_col}{(str) - the column to run the changepoint analysis on. This
can be any column, but for DMR analysis we recommend using the z scores for
a fixed effect variable.}
}
\value{
everything (df) - data frame containing the mean changepoint value for the
\code{z_col} column

everything (df) - data frame containing the mean changepoint value for the
\code{z_col} column
}
\description{
Purpose: run the changepoint analysis for each gene and each cytosine
context

Purpose: run the changepoint analysis for each gene and each cytosine
context
}
