% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sound_DMR.R
\name{create_cols_for_individuals}
\alias{create_cols_for_individuals}
\title{Building Individual Methylation Columns}
\usage{
create_cols_for_individuals(
  Exp_ID_Treated,
  Output_Frame,
  GenePercentPlant,
  GeneDepthPlant,
  GenePercentGroup,
  control = "C"
)
}
\arguments{
\item{Exp_ID_Treated}{(df) - containing the experimental design}

\item{Output_Frame}{(df) - the Output_Frame}

\item{GenePercentPlant}{(df) - summary dataframe built from \code{create_gene_percent_x()}
containing \% methylation for each Plant}

\item{GeneDepthPlant}{(df) - summary dataframe containing read depth for individuals}

\item{GenePercentGroup}{(df) - summary dataframe built from \code{create_gene_percent_x()}
containing \% methylation for groups}

\item{control}{(str) - the control variable in GenePercentGroup}
}
\value{
Output_Frame (df) - the updated Output_Frame with 3 new columns for each individual in
the \code{Exp_IF_Treated} parameter
}
\description{
Set up the output frame to include individual methylation change and Read
Depth columns
}
