% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sound_DMR.R
\name{create_dmr_obj}
\alias{create_dmr_obj}
\title{Create DMR Object}
\usage{
create_dmr_obj(ZoomFrame = dataframe, experimental_design_df = dataframe)
}
\arguments{
\item{ZoomFrame}{(df) containing the input methylation data}

\item{experimental_design_df}{(df) containing the experimental design}

\item{colnames_of_interest}{\emph{optional} (list of strings) - the columns to keep
in the analysis}
}
\value{
out (list) - is a list of dataframes. This is the dmr object.
}
\description{
A function to clean the input data of methylation data. It filters the data
frame to contain only the columns of interest passed in as an argument and
removes rows where the mean methylation is 0.
}
