% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sound_DMR.R
\name{create_formula}
\alias{create_formula}
\title{Create Formula}
\usage{
create_formula(fixed, random)
}
\arguments{
\item{fixed}{(str) - the fixed effects variables. Note: this
value \strong{cannot} be 'ID'. ID is used to merge data together downstream for
running the model so including the value in the effects will break the model.}

\item{random}{(str) - the random effects variables. Note: this
value \strong{cannot} be 'ID'. ID is used to merge data together downstream for
running the model so including the value in the effects will break the model.}
}
\value{
effects_formula (formula) - a formula of the mixed effects
}
\description{
Combine the fixed and random effects into the formula to be used
in the model
}
\examples{
# Create formula of independent and random effects
create_formula(fixed = c('Group'), random = c('Individual'))
# [1] "cbind(Meth, UnMeth) ~ Group + (1 | Individual)"

create_formula(fixed = c('Group', 'Individual'), random = c('Plant * Gene'))
# [1] "cbind(Meth, UnMeth) ~ Group + Individual + (1 | Plant * Gene)"
}
