% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sound_DMR.R
\name{create_methyl_summary}
\alias{create_methyl_summary}
\title{Create Methyl Summary}
\usage{
create_methyl_summary(
  dmr_obj,
  control = "C",
  treated = "T",
  colnames_of_interest = c("Chromosome", "Gene", "Position", "Strand", "CX",
    "Zeroth_pos", "Plant"),
  additional_summary_cols = list()
)
}
\arguments{
\item{dmr_obj}{the dmr object containing the experimental design and raw data}

\item{control}{(str) - the control variable in GenePercentGroup}

\item{treated}{the string representing the treated group}

\item{additional_summary_cols}{a nested list of parameters to create additional
summary columns. Each nested list will be a tuple where the first value is
the summary stats function (e.g. sd, mean, var) and the second value is the
string name of the column on which to run the summary statistic function}
}
\description{
This function creates the methyl_summary that the group and individual DMR
functions rely on for their analysis. It includes some quality control checks
to make sure the output is in the correct format.
}
