% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sound_DMR.R
\name{create_random_effects}
\alias{create_random_effects}
\title{Create Random Effects}
\usage{
create_random_effects(random = c("Group", "ID"))
}
\arguments{
\item{random}{(list of strings) - the random effects elements}
}
\value{
random_effects (str) - the random effects properly formatted
}
\description{
Create a string combining the list of random effects to be passed into the
\code{create_formula()} function
}
\examples{
# Create independent random effects
create_random_effects(c("Group", "Individual"))
# [1] "(1 | Group) + (1 | Individual)"

# Create random effects with an interaction
create_random_effects(c("Group * Individual"))
# [1] "(1 | Group * Individual)"

}
