% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sound_DMR.R
\name{find_changepoint_col_options}
\alias{find_changepoint_col_options}
\title{Find the Columns for Changepoint Analysis}
\usage{
find_changepoint_col_options(DMR_output, Output_Frame = Output_Frame)
}
\arguments{
\item{DMR_output}{(df) - the DMR Analysis output}
}
\value{
z_cols (list of strings) - a list of strings representing the potential
column names that are suggested to be used in the changepoint analysis.
}
\description{
This function is used to find the column names on which to run changepoint
analysis from the \code{group_DMR()} or \code{individual_DMR()} output frames.
These columns are usually the Z score column created from the \code{run_model}
model summary. The Z score column is based on the fixed effects that are
passed into the model. Note: any column can be passed into the changepoint
analysis function, the output of this function are the columns deemed to
provide the greatest use to the analysis.
}
