% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sound_DMR.R
\name{find_cpt_mean}
\alias{find_cpt_mean}
\title{Find CPT Mean}
\usage{
find_cpt_mean(data, z_col, penalty)
}
\arguments{
\item{data}{(df) - the dataframe that has been subsetted for the gene and cytosine
context}

\item{z_col}{(str) - the column to run the changepoint analysis on. This
can be any column, but for DMR analysis we recommend using the z scores for
a fixed effect variable.}

\item{penalty}{(int) - the penalty value to include for the changepoint analysis}
}
\value{
changepoint_object (S4) - containing the changepoint information
}
\description{
Purpose: find the changepoints based on mean values and include error catches
}
