% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sound_DMR.R
\name{find_DMR}
\alias{find_DMR}
\title{DMR Analysis}
\usage{
find_DMR(
  methyl_sum,
  dmr_obj,
  fixed = c("Group"),
  random = c("Plant"),
  reads_threshold = 3,
  model,
  control = "",
  analysis_type
)
}
\arguments{
\item{methyl_sum}{(df) - the read depth and methylation
change information}

\item{dmr_obj}{(list) - the dmr_object containing the experimental design df and zoomFrame_filtered}

\item{fixed}{(str) - the fixed effects variables. Note: this
value \strong{cannot} be 'ID'. ID is used to merge data together downstream for
running the model so including the value in the effects will break the model.}

\item{random}{(str) - the random effects variables. Note: this
value \strong{cannot} be 'ID'. ID is used to merge data together downstream for
running the model so including the value in the effects will break the model.}

\item{reads_threshold}{(int) - the number of reads that are each
methylated and unmethylated. This is important since data containing only one
methylated read is unlikely to provide statistical power to our analysis. A
value of 3 here means that the read depth is \emph{at least} the depth provided by
the threshold. Both the methylated and unmethylated samples need a read depth
greater than or equal to this threshold in order to be considered for the model.}

\item{control}{(str) - the control variable}

\item{analysis_type}{(str) - either "individual" or "group"}
}
\value{
Output_Frame (df)
}
\description{
This function runs the DMR analysis based on whether analysis is between
groups or for an individual compared to a group.
}
