% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sound_DMR.R
\name{generate_megaframe}
\alias{generate_megaframe}
\title{generate_megaframe}
\usage{
generate_megaframe(
  methyl_bed_list = All_methyl_beds,
  Sample_count,
  Methyl_call_type = "Dorado",
  File_prefix = "",
  max_read_depth = 100
)
}
\arguments{
\item{methyl_bed_list}{(list) - ONT methyl bed filenames for each individual contained within the directory. This will just be a list of bedfile names.
Hint : The input will be the "methyl_bed_list" vector that you create in the previous step.}

\item{Sample_count}{(int) - This is required to assign proper alphabet codes. If you need to include the samples from a previous round, then enter the total number of samples from the previous round here. Default is 0. By default alphabetizing starts with 'A'.}

\item{Methyl_call_type}{(str) - A string that includes information about the type of run. Currently this package works on Megalodon , DSP (DeepSignal Plant) and Bonito.}

\item{File_prefix}{(Flexible str) - This is to add a prefix to all the files that get exported and saved to the working directory while running the function.}
}
\value{
Megaframe(df) - Clean data frame containing combined methyl bed information for every individual in the experiment.
}
\description{
A function to create a single-combined data frame from individual methyl beds in the experiment
This function only works with bedfiles output from only either of the three methylation call algorithms - DeepSignal Plant(DSP), Megalodon and Bonito.
}
\examples{
# Basic usage for methyl_call_type
# generate_megaframe(Methyl_call_type="DSP") # OR
# get_standard_methyl_bed(Methyl_call_type="Megalodon") # OR
# get_standard_methyl_bed(Methyl_call_type="Bonito")
}
