% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sound_DMR.R
\name{generate_methylframe}
\alias{generate_methylframe}
\title{Create Methylframe}
\usage{
generate_methylframe(
  methyl_bed_list = All_methyl_beds,
  Sample_count = 0,
  Methyl_call_type = "Dorado",
  filter_NAs = 0,
  max_read_depth = 100,
  gene_info = FALSE,
  gene_coordinate_file = NULL,
  Gene_column = "",
  target_info = TRUE,
  gene_list = gene_coordinate_file[[Gene_column]],
  File_prefix = "Sample"
)
}
\arguments{
\item{methyl_bed_list}{(list) - ONT methyl bed filenames for each individual contained within the directory. This will just be a list of bedfile names. The input will be the "methyl_bed_list" vector that you create in the previous step.}

\item{Sample_count}{(int) - This is required to assign proper alphabet codes. If you need to include the samples from a previous round, then enter the total number of samples from the previous round here. Default is 0. By default alphabetizing starts with 'A'.}

\item{Methyl_call_type}{(str) - A string that includes information about the type of run. Currently this package works on Megalodon , DSP (DeepSignal Plant) and Bonito.}

\item{filter_NAs}{(int) - Select this parameter based on the histogram plot. This will filter out NAs based on per sample, by default this is 0.}

\item{gene_info}{(str) - This takes in a boolean variable. If the gene info- coordinates, gene name etc are present then make sure to have the gene-cordinates.csv file in the right format (as shown in the sample data on github).}

\item{target_info}{(Boolean) - This takes in TRUE or FALSE. Enter TRUE only if the megaframe contains target genes that need to be differentiated from non-targets.}

\item{gene_list}{(list) - Provide a list of target genes to distiguish from non-target genes within the Zoomframe. By deafult it will take in the All the genes from the gene coordinates file.}

\item{File_prefix}{(Flexible str) - This is to add a prefix to all the files that get exported and saved to the working directory while running the function.}

\item{gene_cordinate_file}{(str) - File containing gene-coordinate info. It is important to be in a specific format and should have but not limited to the following columns : Chromosome | Gene_Name | Low | High | Adapt_Low | Adapt_High
Low and High : These are Start and Stop cordinates of the gene. Low is always the lower coordinate which could be start for the positive stranded gene and stop for the negative stranded gene and vice versa.
Adapt_Low and Adapt_High : Cordinates for adaptive regions around the Lower and Higher co-ordinate of the gene respectively}
}
\value{
Megaframe(df) or Zoomframe(df) - Clean data frame containing combined methyl bed information for every individual in the experiment.
}
\description{
A function to create either Megaframe (A single data frame containing all samples bedMethyl data) or Zoomframe (A version of the Megaframe which includes the Gene Name, Position Zero'ed in on ATG of the respective gene and other columns to help with downstream analysis)
NOTE:
\enumerate{
\item In both cases, Megaframe data will be exported in the current working directory and Zoomframe will only be exported if the Gene coordinates file is provided
\item If this function is run using 'gene_info= TRUE' then it needs additional parameters like gene_cordinate_file,  Gene_column,target_info, gene_list etc to precoeed. If no such file is present, enter gene_info as false and
}
}
