% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sound_DMR.R
\name{generate_zoomframe}
\alias{generate_zoomframe}
\title{Create Zoomframe}
\usage{
generate_zoomframe(
  gene_cord_df,
  MFrame,
  Gene_col,
  target_info = TRUE,
  gene_list = gene_cord_df[[Gene_col]],
  File_prefix = ""
)
}
\arguments{
\item{gene_cord_df}{(df) - Data frame containing gene-coordinate info}

\item{MFrame}{(df) - Megaframe data from the previous function}

\item{Gene_col}{(str) - Use this column to specifify wether to add Gene Names or Ids in the Zoomframe.}

\item{target_info}{(Boolean) - This takes in TRUE or FALSE. Enter TRUE only if the megaframe contains target genes that need to be differentiated from non-targets.}

\item{gene_list}{(list) - Provide a list of target genes to distiguish from non-target genes within the Zoomframe. By deafult it will take in the All the genes from the gene coordinates file.}

\item{File_prefix}{(Flexible str) - This is to add a prefix to all the files that get exported while running the function.}

\item{filter_NAs}{(int) - Select this parameter based on the histogram plot. This will filter out NAs based on per sample}
}
\value{
Zoomframe (df) - Similar to Megaframe except this includes more information on targets, positions zero'ed to ATG for each target and a few other information.
}
\description{
A function to create a single-merged data frame from individual methyl beds in the experiment
}
