% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sound_DMR.R
\name{get_standard_methyl_bed}
\alias{get_standard_methyl_bed}
\title{get_standard_methyl_bed}
\usage{
get_standard_methyl_bed(
  Methyl_bed = "Methyl.bed",
  Sample_ID = "S1",
  Methyl_call_type = "Dorado"
)
}
\arguments{
\item{Methyl_bed}{(df) - Data frame containing the ONT methylation calls in a bed file for each individual separately}

\item{Sample_ID}{(str) - A string that is used inplace of sample name to keep it uniform. We are using an enumerator to generate this based on the number of samples/individuals in the experiment.}

\item{Methyl_call_type}{(str) - A string that includes information about the type of run. Currently this package works on Megalodon , DSP (DeepSignal Plant), Dorado and Bonito. Default call type is Dorado.}
}
\value{
Methyl_bed_sub (df) - Standard data frame methyl file for every individual with Meth, Unmeth and Per_Meth columns
}
\description{
A function to create a data frame for every individual in the experimental design without having to re run for every individual separately.
This function takes in the methyl_bed file, subsets and then creates Methylated and unmethylated counts for each position to be used in the next steps.
}
