% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sound_DMR.R, R/Sound_DMR_formatted.R
\name{group_DMR}
\alias{group_DMR}
\title{Group DMR Analysis}
\usage{
group_DMR(
  Output_Frame,
  zoom_frame_filtered,
  experimental_design_df,
  fixed = c("Group"),
  random = c("Plant"),
  reads_threshold = 3,
  model = "binomial",
  colnames_of_interest = c("Chromosome", "Gene", "Position", "Strand", "CX",
    "Zeroth_pos", "Individual")
)

group_DMR(
  Output_Frame,
  zoom_frame_filtered,
  experimental_design_df,
  fixed = c("Group"),
  random = c("Plant"),
  reads_threshold = 3,
  model = "binomial",
  colnames_of_interest = c("Chromosome", "Gene", "Position", "Strand", "CX",
    "Zeroth_pos", "Individual")
)
}
\arguments{
\item{Output_Frame}{(df) - the read depth and methylation
change information}

\item{zoom_frame_filtered}{(df) - the percent methylation information}

\item{experimental_design_df}{(df) - the experimental design}

\item{fixed}{(str) - the fixed effects variables. Note: this
value \strong{cannot} be 'ID'. ID is used to merge data together downstream for
running the model so including the value in the effects will break the model.}

\item{random}{(str) - the random effects variables. Note: this
value \strong{cannot} be 'ID'. ID is used to merge data together downstream for
running the model so including the value in the effects will break the model.}

\item{reads_threshold}{(int) - the number of reads that are each
methylated and unmethylated. This is important since data containing only one
methylated read is unlikely to provide statistical power to our analysis. A
value of 3 here means that the read depth is \emph{at least} the depth provided by
the threshold. Both the methylated and unmethylated samples need a read depth
greater than or equal to this threshold in order to be considered for the model.}

\item{colnames_of_interest}{\emph{optional} (list of strings) - the columns to
keep in the analysis}

\item{ZoomFrame_filtered}{(df) - the percent methylation information}
}
\value{
Output_Frame (df) - Output_Frame with the summary statistics from the model

Output_Frame (df) - Output_Frame with the summary statistics from the model
}
\description{
To run a binomial model to compare the methylation between groups

To run a binomial model to compare the methylation between groups
}
