% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sound_DMR.R, R/Sound_DMR_formatted.R
\name{pivot_and_subset}
\alias{pivot_and_subset}
\title{Arrange Data into Long Format and Subset}
\usage{
pivot_and_subset(
  data,
  starts_with_cols = "start",
  values_to,
  colnames_of_interest = c("Chromosome", "Gene", "Position", "Strand", "CX",
    "Zeroth_pos", "Individual")
)

pivot_and_subset(
  data,
  starts_with_cols = "start",
  values_to,
  colnames_of_interest = c("Chromosome", "Gene", "Position", "Strand", "CX",
    "Zeroth_pos", "Individual")
)
}
\arguments{
\item{data}{(df) - the \code{zoom_frame_filtered}}

\item{starts_with_cols}{(str) - the string pattern that the columns start with}

\item{values_to}{A string specifying the name of the column to create
from the data stored in cell values. If \code{names_to} is a character
containing the special \code{.value} sentinel, this value will be ignored,
and the name of the value column will be derived from part of the
existing column names.}

\item{colnames_of_interest}{\emph{optional} (list of strings) - the columns to
keep in the analysis}
}
\value{
pivoted_and_subsetted_df (df) - the cleaned data frame that has been
pivoted to be longer and subsetted to only include the important columns.

pivoted_and_subsetted_df (df) - the cleaned data frame that has been
pivoted to be longer and subsetted to only include the important columns.
}
\description{
This function is used to get data into LongPercent and LongMeth formats to
allow for summary analysis. This function utilizes \code{tidyr::pivot_longer()}
and \code{subset_cols()} from this package.

This function is used to get data into long_percent and long_meth formats to
allow for summary analysis. This function utilizes \code{tidyr::pivot_longer()}
and \code{subset_cols()} from this package.
}
