% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sound_DMR.R
\name{plot_changepoints}
\alias{plot_changepoints}
\title{Plot Changepoints}
\usage{
plot_changepoints(
  data,
  changepoint_obj,
  gene_name,
  penalty_val,
  cyt_context,
  z_col
)
}
\arguments{
\item{data}{(df) - containing the cytosines with their percent methylation}

\item{changepoint_obj}{(S4) - the S4 object created from finding the changepoints}

\item{gene_name}{(str) - the name of the gene}

\item{penalty_val}{(int) - the penalty value being used}

\item{cyt_context}{(str) - the cytosine context}

\item{z_col}{(str) - the column upon which the changepoints are calculated}
}
\value{
changepoint_plot (plot) - the plot of the changepoints
}
\description{
Plot the changepoint data
}
