% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sound_DMR.R, R/Sound_DMR_formatted.R
\name{run_model}
\alias{run_model}
\title{Run Model}
\usage{
run_model(data, i, Output_Frame, formula, model_type, individual_name_z = "")

run_model(data, i, Output_Frame, formula, model_type, individual_name_z = "")
}
\arguments{
\item{data}{(df) - the information to put in the model, usually this is
in 'long' format}

\item{i}{(int) - the row number}

\item{Output_Frame}{(df) - all the summary information}

\item{formula}{(formula) - the formula to use in the model}

\item{individual_name_z}{(str) - \emph{optional} the name of the z column when running
individual DMR analysis}
}
\value{
Output_Frame (df) - updated with the summary data from the model

Output_Frame (df) - updated with the summary data from the model
}
\description{
Function to run the correct model

Function to run the correct model
}
