% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sound_DMR.R
\name{save_model_summary}
\alias{save_model_summary}
\title{Save Model Summary}
\usage{
save_model_summary(i, Output_Frame, model_summary, ind_name = "")
}
\arguments{
\item{i}{(int) - the row number of the dataframe}

\item{Output_Frame}{(df) - all the information}

\item{model_summary}{(df) - the summary statistics from the model}

\item{ind_name}{(str) - \emph{optional} containing the name of the z score column for
individual DMR analysis}
}
\value{
Output_Frame (df) - the data frame with the proper row updated to include
the summary statistics
}
\description{
Update the \code{Output_Frame} dataframe for a given row with the summary statistics
provided by the model
}
