% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sound_DMR.R
\name{save_plot}
\alias{save_plot}
\title{Save Plot}
\usage{
save_plot(file_name, plot_obj)
}
\arguments{
\item{file_name}{(str) - the name of the filepath}

\item{plot_obj}{(obj) - the plot object to save}
}
\description{
Save a plot
}
