% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sound_DMR.R
\name{sound_score}
\alias{sound_score}
\title{Sound Score}
\usage{
sound_score(
  changepoint_OF = dataframe,
  Statistic = "Z_GroupT_small",
  Per_Change = "Treat_V_Control",
  Control = "Control",
  other_columns = c("Estimate_GroupT_small"),
  CF = FALSE,
  UserFunction = NA
)
}
\arguments{
\item{changepoint_OF}{(df) - the OF file that has had changepoint calling ran on it}

\item{Statistic}{(str) - name of the test statistic that changepoint was run on}

\item{Per_Change}{(str) - Column that captures the change in methylation between two groups of interest}

\item{other_columns}{(str) - Names of any other columns that one is interested in calculating region averaqges on.}

\item{CF}{(str) - True/False if there is a custom function to be calculated in addition to soundscore}

\item{User_Function}{(str) - The function to use for the custome function}
}
\value{
Aggregated_Changepoint_Object (list) - File that contains information about methylation patterns within each changepoint region
}
\description{
Takes in an Output_Frame that has been broken in to changepoint regions based on a specific test statistic of interest, and creates an aggregated changepoint region file that includes summary statistics for each region, and the "sound score" a measure of the strength of a DMR within that region.
}
