% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sound_DMR.R
\name{split_by_chromosome}
\alias{split_by_chromosome}
\title{split_by_chromosome}
\usage{
split_by_chromosome(input_file)
}
\arguments{
\item{input_file}{(str) - A string with the name of the input file. It uses a
fill (absolute) path.}
}
\value{
output_filelist(list) - A list with the bedfile names of the output files.
}
\description{
A function to split a bed file into multiple file by chromosome.
It will create a new directory per each chromosome.
}
