% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sound_DMR.R
\name{split_by_chunk}
\alias{split_by_chunk}
\title{split_by_chunk}
\usage{
split_by_chunk(input_file, chunk_size, output_dir = "./chunks/")
}
\arguments{
\item{input_file}{(str) - A string with the name of the input file.}

\item{chunk_size}{(int) - Maximum region/position in the bedfile per chunk.}
}
\description{
A function to split a bed file into multiple chunks.
}
