% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sound_DMR.R, R/Sound_DMR_formatted.R
\name{subset_cols}
\alias{subset_cols}
\title{Subset Data Frames to Contain Only Columns of Interest}
\usage{
subset_cols(
  df,
  colnames_of_interest = c("Chromosome", "Gene", "Position", "Strand", "CX",
    "Zeroth_pos", "Individual")
)

subset_cols(
  df,
  colnames_of_interest = c("Chromosome", "Gene", "Position", "Strand", "CX",
    "Zeroth_pos", "Individual")
)
}
\arguments{
\item{df}{(df) data frame to subset}

\item{colnames_of_interest}{\emph{optional} (list of strings) - the columns to
keep in the analysis}
}
\value{
subsetted_df (df) - data frame with only the columns of interest

subsetted_df (df) - data frame with only the columns of interest
}
\description{
Keep only the necessary columns from a given data frame

Keep only the necessary columns from a given data frame
}
\examples{
# Generate data to subset
df <- data.frame(
 x = c(1, 2, 3, 4),
 y = c(6, 7, 8, 9),
 z = c(14, 23, 15, 4))

# Subset the data
subset_cols(df, c('x', 'z'))

# Generate data to subset
df <- data.frame(
 x = c(1, 2, 3, 4),
 y = c(6, 7, 8, 9),
 z = c(14, 23, 15, 4))

# Subset the data
subset_cols(df, c('x', 'z'))

}
