% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sound_DMR.R, R/Sound_DMR_formatted.R
\name{subset_methyl_summary}
\alias{subset_methyl_summary}
\title{Subset Methyl Summary}
\usage{
subset_methyl_summary(methyl_summary, individuals_to_keep)

subset_methyl_summary(methyl_summary, individuals_to_keep)
}
\arguments{
\item{methyl_summary}{(df) - the input data}

\item{individuals_to_keep}{(list of strings) - the individuals to keep for
analysis}
}
\value{
methyl_summary_subset (df) - the data only including the individuals
for analysis

methyl_summary_subset (df) - the data only including the individuals
for analysis
}
\description{
A function to subset the methyl_summary file to only include individuals of
interest in the analysis

A function to subset the methyl_summary file to only include individuals of
interest in the analysis
}
