% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sound_DMR.R
\name{run_model}
\alias{run_model}
\title{Run Model}
\usage{
run_model(data, i, input_frame, formula, model_type, individual_name_z = "")
}
\arguments{
\item{data}{(df) - the information to put in the model, usually this is
in "long" format}

\item{i}{(int) - the row number}

\item{formula}{(formula) - the formula to use in the model}

\item{individual_name_z}{(str) - \emph{optional} the name of the z column when running
individual DMR analysis}
}
\value{
input_frame (df) - updated with the summary data from the model
}
\description{
Function to run the correct model
}
